fix(execution-view): move constants and types out of "use server" module to unblock build
Next.js only allows async function exports from a "use server" file. TEAM_RAM_AGENTS
and the shared types now live in execution-view.types.ts (a plain module); the server
action, client component, and tests import from there. No API/response-shape change.

Fixes: build, analyze, MCP Browser Tests (next build page-data collection).
